--- a/mapstructure.go
+++ b/mapstructure.go
@@ -36,6 +36,10 @@ type DecodeHookFunc interface{}
 // the source and target types.
 type DecodeHookFuncType func(reflect.Type, reflect.Type, interface{}) (interface{}, error)
 
+// DecodeHookFuncTypeContext is a DecodeHookFunc which has complete information about
+// the source and target types and additional contextual information about the field.
+type DecodeHookFuncTypeContext func(reflect.Type, reflect.Type, interface{}, *DecodeContext) (interface{}, error)
+
 // DecodeHookFuncKind is a DecodeHookFunc which knows only the Kinds of the
 // source and target types.
 type DecodeHookFuncKind func(reflect.Kind, reflect.Kind, interface{}) (interface{}, error)
@@ -113,6 +117,18 @@ type Metadata struct {
 	// weren't decoded since there was no matching field in the result interface
 	Unused []string
 }
+
+// DecodeContext contextual information about the current decode operation.
+type DecodeContext struct {
+	Name  string
+	IsKey bool
+}
+
+func (d DecodeContext) copy() *DecodeContext {
+	return &d
+}
+
+var noContext = &DecodeContext{}
 
 // Decode takes an input structure and uses reflection to translate it to
 // the output structure. output must be a pointer to a map or struct.
@@ -219,19 +235,19 @@ func NewDecoder(config *DecoderConfig) (*Decoder, error) {
 // Decode decodes the given raw interface to the target pointer specified
 // by the configuration.
 func (d *Decoder) Decode(input interface{}) error {
-	return d.decode("", input, reflect.ValueOf(d.config.Result).Elem())
+	return d.decode(noContext, input, reflect.ValueOf(d.config.Result).Elem())
 }
 
 // Decodes an unknown data type into a specific reflection value.
-func (d *Decoder) decode(name string, input interface{}, outVal reflect.Value) error {
+func (d *Decoder) decode(ctx *DecodeContext, input interface{}, outVal reflect.Value) error {
 	if input == nil {
 		// If the data is nil, then we don't set anything, unless ZeroFields is set
 		// to true.
 		if d.config.ZeroFields {
 			outVal.Set(reflect.Zero(outVal.Type()))
 
-			if d.config.Metadata != nil && name != "" {
-				d.config.Metadata.Keys = append(d.config.Metadata.Keys, name)
+			if d.config.Metadata != nil && ctx.Name != "" {
+				d.config.Metadata.Keys = append(d.config.Metadata.Keys, ctx.Name)
 			}
 		}
 		return nil
@@ -242,8 +258,8 @@ func (d *Decoder) decode(name string, input interface{}, outVal reflect.Value) e
 		// If the input value is invalid, then we just set the value
 		// to be the zero value.
 		outVal.Set(reflect.Zero(outVal.Type()))
-		if d.config.Metadata != nil && name != "" {
-			d.config.Metadata.Keys = append(d.config.Metadata.Keys, name)
+		if d.config.Metadata != nil && ctx.Name != "" {
+			d.config.Metadata.Keys = append(d.config.Metadata.Keys, ctx.Name)
 		}
 		return nil
 	}
@@ -253,9 +269,9 @@ func (d *Decoder) decode(name string, input interface{}, outVal reflect.Value) e
 		var err error
 		input, err = DecodeHookExec(
 			d.config.DecodeHook,
-			inputVal.Type(), outVal.Type(), input)
+			inputVal.Type(), outVal.Type(), input, ctx)
 		if err != nil {
-			return fmt.Errorf("error decoding '%s': %s", name, err)
+			return fmt.Errorf("error decoding '%s': %s", ctx.Name, err)
 		}
 	}
 
@@ -263,38 +279,38 @@ func (d *Decoder) decode(name string, input interface{}, outVal reflect.Value) e
 	inputKind := getKind(outVal)
 	switch inputKind {
 	case reflect.Bool:
-		err = d.decodeBool(name, input, outVal)
+		err = d.decodeBool(ctx.Name, input, outVal)
 	case reflect.Interface:
-		err = d.decodeBasic(name, input, outVal)
+		err = d.decodeBasic(ctx, input, outVal)
 	case reflect.String:
-		err = d.decodeString(name, input, outVal)
+		err = d.decodeString(ctx.Name, input, outVal)
 	case reflect.Int:
-		err = d.decodeInt(name, input, outVal)
+		err = d.decodeInt(ctx.Name, input, outVal)
 	case reflect.Uint:
-		err = d.decodeUint(name, input, outVal)
+		err = d.decodeUint(ctx.Name, input, outVal)
 	case reflect.Float32:
-		err = d.decodeFloat(name, input, outVal)
+		err = d.decodeFloat(ctx.Name, input, outVal)
 	case reflect.Struct:
-		err = d.decodeStruct(name, input, outVal)
+		err = d.decodeStruct(ctx, input, outVal)
 	case reflect.Map:
-		err = d.decodeMap(name, input, outVal)
+		err = d.decodeMap(ctx, input, outVal)
 	case reflect.Ptr:
-		err = d.decodePtr(name, input, outVal)
+		err = d.decodePtr(ctx, input, outVal)
 	case reflect.Slice:
-		err = d.decodeSlice(name, input, outVal)
+		err = d.decodeSlice(ctx, input, outVal)
 	case reflect.Array:
-		err = d.decodeArray(name, input, outVal)
+		err = d.decodeArray(ctx, input, outVal)
 	case reflect.Func:
-		err = d.decodeFunc(name, input, outVal)
+		err = d.decodeFunc(ctx.Name, input, outVal)
 	default:
 		// If we reached this point then we weren't able to decode it
-		return fmt.Errorf("%s: unsupported type: %s", name, inputKind)
+		return fmt.Errorf("%s: unsupported type: %s", ctx.Name, inputKind)
 	}
 
 	// If we reached here, then we successfully decoded SOMETHING, so
 	// mark the key as used if we're tracking metainput.
-	if d.config.Metadata != nil && name != "" {
-		d.config.Metadata.Keys = append(d.config.Metadata.Keys, name)
+	if d.config.Metadata != nil && ctx.Name != "" {
+		d.config.Metadata.Keys = append(d.config.Metadata.Keys, ctx.Name)
 	}
 
 	return err
@@ -302,9 +318,9 @@ func (d *Decoder) decode(name string, input interface{}, outVal reflect.Value) e
 
 // This decodes a basic type (bool, int, string, etc.) and sets the
 // value to "data" of that type.
-func (d *Decoder) decodeBasic(name string, data interface{}, val reflect.Value) error {
+func (d *Decoder) decodeBasic(ctx *DecodeContext, data interface{}, val reflect.Value) error {
 	if val.IsValid() && val.Elem().IsValid() {
-		return d.decode(name, data, val.Elem())
+		return d.decode(ctx, data, val.Elem())
 	}
 	dataVal := reflect.ValueOf(data)
 	if !dataVal.IsValid() {
@@ -315,7 +331,7 @@ func (d *Decoder) decodeBasic(name string, data interface{}, val reflect.Value) 
 	if !dataValType.AssignableTo(val.Type()) {
 		return fmt.Errorf(
 			"'%s' expected type '%s', got '%s'",
-			name, val.Type(), dataValType)
+			ctx.Name, val.Type(), dataValType)
 	}
 
 	val.Set(dataVal)
@@ -532,7 +548,7 @@ func (d *Decoder) decodeFloat(name string, data interface{}, val reflect.Value) 
 	return nil
 }
 
-func (d *Decoder) decodeMap(name string, data interface{}, val reflect.Value) error {
+func (d *Decoder) decodeMap(ctx *DecodeContext, data interface{}, val reflect.Value) error {
 	valType := val.Type()
 	valKeyType := valType.Key()
 	valElemType := valType.Elem()
@@ -551,33 +567,36 @@ func (d *Decoder) decodeMap(name string, data interface{}, val reflect.Value) er
 	dataVal := reflect.Indirect(reflect.ValueOf(data))
 	switch dataVal.Kind() {
 	case reflect.Map:
-		return d.decodeMapFromMap(name, dataVal, val, valMap)
+		return d.decodeMapFromMap(ctx, dataVal, val, valMap)
 
 	case reflect.Struct:
-		return d.decodeMapFromStruct(name, dataVal, val, valMap)
+		return d.decodeMapFromStruct(ctx, dataVal, val, valMap)
 
 	case reflect.Array, reflect.Slice:
 		if d.config.WeaklyTypedInput {
-			return d.decodeMapFromSlice(name, dataVal, val, valMap)
+			return d.decodeMapFromSlice(ctx, dataVal, val, valMap)
 		}
 
 		fallthrough
 
 	default:
-		return fmt.Errorf("'%s' expected a map, got '%s'", name, dataVal.Kind())
+		return fmt.Errorf("'%s' expected a map, got '%s'", ctx.Name, dataVal.Kind())
 	}
 }
 
-func (d *Decoder) decodeMapFromSlice(name string, dataVal reflect.Value, val reflect.Value, valMap reflect.Value) error {
+func (d *Decoder) decodeMapFromSlice(ctx *DecodeContext, dataVal reflect.Value, val reflect.Value, valMap reflect.Value) error {
 	// Special case for BC reasons (covered by tests)
 	if dataVal.Len() == 0 {
 		val.Set(valMap)
 		return nil
 	}
 
+	ctxc := ctx.copy()
+	name := ctxc.Name
 	for i := 0; i < dataVal.Len(); i++ {
+		ctxc.Name = fmt.Sprintf("%s[%d]", name, i)
 		err := d.decode(
-			fmt.Sprintf("%s[%d]", name, i),
+			ctxc,
 			dataVal.Index(i).Interface(), val)
 		if err != nil {
 			return err
@@ -587,7 +606,7 @@ func (d *Decoder) decodeMapFromSlice(name string, dataVal reflect.Value, val ref
 	return nil
 }
 
-func (d *Decoder) decodeMapFromMap(name string, dataVal reflect.Value, val reflect.Value, valMap reflect.Value) error {
+func (d *Decoder) decodeMapFromMap(ctx *DecodeContext, dataVal reflect.Value, val reflect.Value, valMap reflect.Value) error {
 	valType := val.Type()
 	valKeyType := valType.Key()
 	valElemType := valType.Elem()
@@ -595,20 +614,24 @@ func (d *Decoder) decodeMapFromMap(name string, dataVal reflect.Value, val refle
 	// Accumulate errors
 	errors := make([]string, 0)
 
+	ctxc := ctx.copy()
+	name := ctxc.Name
 	for _, k := range dataVal.MapKeys() {
-		fieldName := fmt.Sprintf("%s[%s]", name, k)
+		ctxc.Name = fmt.Sprintf("%s[%s]", name, k)
+		ctxc.IsKey = true
 
 		// First decode the key into the proper type
 		currentKey := reflect.Indirect(reflect.New(valKeyType))
-		if err := d.decode(fieldName, k.Interface(), currentKey); err != nil {
+		if err := d.decode(ctxc, k.Interface(), currentKey); err != nil {
 			errors = appendErrors(errors, err)
 			continue
 		}
 
 		// Next decode the data into the proper type
+		ctxc.IsKey = false
 		v := dataVal.MapIndex(k).Interface()
 		currentVal := reflect.Indirect(reflect.New(valElemType))
-		if err := d.decode(fieldName, v, currentVal); err != nil {
+		if err := d.decode(ctxc, v, currentVal); err != nil {
 			errors = appendErrors(errors, err)
 			continue
 		}
@@ -627,8 +650,9 @@ func (d *Decoder) decodeMapFromMap(name string, dataVal reflect.Value, val refle
 	return nil
 }
 
-func (d *Decoder) decodeMapFromStruct(name string, dataVal reflect.Value, val reflect.Value, valMap reflect.Value) error {
+func (d *Decoder) decodeMapFromStruct(ctx *DecodeContext, dataVal reflect.Value, val reflect.Value, valMap reflect.Value) error {
 	typ := dataVal.Type()
+	ctxc := ctx.copy()
 	for i := 0; i < typ.NumField(); i++ {
 		// Get the StructField first since this is a cheap operation. If the
 		// field is unexported, then ignore it.
@@ -680,7 +704,8 @@ func (d *Decoder) decodeMapFromStruct(name string, dataVal reflect.Value, val re
 			mType := reflect.MapOf(vKeyType, vElemType)
 			vMap := reflect.MakeMap(mType)
 
-			err := d.decode(keyName, x.Interface(), vMap)
+			ctxc.Name = keyName
+			err := d.decode(ctxc, x.Interface(), vMap)
 			if err != nil {
 				return err
 			}
@@ -705,7 +730,7 @@ func (d *Decoder) decodeMapFromStruct(name string, dataVal reflect.Value, val re
 	return nil
 }
 
-func (d *Decoder) decodePtr(name string, data interface{}, val reflect.Value) error {
+func (d *Decoder) decodePtr(ctx *DecodeContext, data interface{}, val reflect.Value) error {
 	// Create an element of the concrete (non pointer) type and decode
 	// into that. Then set the value of the pointer to this type.
 	valType := val.Type()
@@ -717,13 +742,13 @@ func (d *Decoder) decodePtr(name string, data interface{}, val reflect.Value) er
 			realVal = reflect.New(valElemType)
 		}
 
-		if err := d.decode(name, data, reflect.Indirect(realVal)); err != nil {
+		if err := d.decode(ctx, data, reflect.Indirect(realVal)); err != nil {
 			return err
 		}
 
 		val.Set(realVal)
 	} else {
-		if err := d.decode(name, data, reflect.Indirect(val)); err != nil {
+		if err := d.decode(ctx, data, reflect.Indirect(val)); err != nil {
 			return err
 		}
 	}
@@ -743,7 +768,7 @@ func (d *Decoder) decodeFunc(name string, data interface{}, val reflect.Value) e
 	return nil
 }
 
-func (d *Decoder) decodeSlice(name string, data interface{}, val reflect.Value) error {
+func (d *Decoder) decodeSlice(ctx *DecodeContext, data interface{}, val reflect.Value) error {
 	dataVal := reflect.Indirect(reflect.ValueOf(data))
 	dataValKind := dataVal.Kind()
 	valType := val.Type()
@@ -763,16 +788,16 @@ func (d *Decoder) decodeSlice(name string, data interface{}, val reflect.Value) 
 						return nil
 					}
 				case dataValKind == reflect.String && valElemType.Kind() == reflect.Uint8:
-					return d.decodeSlice(name, []byte(dataVal.String()), val)
+					return d.decodeSlice(ctx, []byte(dataVal.String()), val)
 				// All other types we try to convert to the slice type
 				// and "lift" it into it. i.e. a string becomes a string slice.
 				default:
 					// Just re-try this function with data as a slice.
-					return d.decodeSlice(name, []interface{}{data}, val)
+					return d.decodeSlice(ctx, []interface{}{data}, val)
 				}
 			}
 			return fmt.Errorf(
-				"'%s': source data must be an array or slice, got %s", name, dataValKind)
+				"'%s': source data must be an array or slice, got %s", ctx.Name, dataValKind)
 
 		}
 
@@ -783,6 +808,9 @@ func (d *Decoder) decodeSlice(name string, data interface{}, val reflect.Value) 
 	// Accumulate any errors
 	errors := make([]string, 0)
 
+	ctxc := ctx.copy()
+	name := ctxc.Name
+
 	for i := 0; i < dataVal.Len(); i++ {
 		currentData := dataVal.Index(i).Interface()
 		for valSlice.Len() <= i {
@@ -790,8 +818,8 @@ func (d *Decoder) decodeSlice(name string, data interface{}, val reflect.Value) 
 		}
 		currentField := valSlice.Index(i)
 
-		fieldName := fmt.Sprintf("%s[%d]", name, i)
-		if err := d.decode(fieldName, currentData, currentField); err != nil {
+		ctxc.Name = fmt.Sprintf("%s[%d]", name, i)
+		if err := d.decode(ctxc, currentData, currentField); err != nil {
 			errors = appendErrors(errors, err)
 		}
 	}
@@ -807,7 +835,7 @@ func (d *Decoder) decodeSlice(name string, data interface{}, val reflect.Value) 
 	return nil
 }
 
-func (d *Decoder) decodeArray(name string, data interface{}, val reflect.Value) error {
+func (d *Decoder) decodeArray(ctx *DecodeContext, data interface{}, val reflect.Value) error {
 	dataVal := reflect.Indirect(reflect.ValueOf(data))
 	dataValKind := dataVal.Kind()
 	valType := val.Type()
@@ -832,17 +860,17 @@ func (d *Decoder) decodeArray(name string, data interface{}, val reflect.Value) 
 				// and "lift" it into it. i.e. a string becomes a string array.
 				default:
 					// Just re-try this function with data as a slice.
-					return d.decodeArray(name, []interface{}{data}, val)
+					return d.decodeArray(ctx, []interface{}{data}, val)
 				}
 			}
 
 			return fmt.Errorf(
-				"'%s': source data must be an array or slice, got %s", name, dataValKind)
+				"'%s': source data must be an array or slice, got %s", ctx.Name, dataValKind)
 
 		}
 		if dataVal.Len() > arrayType.Len() {
 			return fmt.Errorf(
-				"'%s': expected source data to have length less or equal to %d, got %d", name, arrayType.Len(), dataVal.Len())
+				"'%s': expected source data to have length less or equal to %d, got %d", ctx.Name, arrayType.Len(), dataVal.Len())
 
 		}
 
@@ -853,12 +881,15 @@ func (d *Decoder) decodeArray(name string, data interface{}, val reflect.Value) 
 	// Accumulate any errors
 	errors := make([]string, 0)
 
+	ctxc := ctx.copy()
+	name := ctxc.Name
+
 	for i := 0; i < dataVal.Len(); i++ {
 		currentData := dataVal.Index(i).Interface()
 		currentField := valArray.Index(i)
 
-		fieldName := fmt.Sprintf("%s[%d]", name, i)
-		if err := d.decode(fieldName, currentData, currentField); err != nil {
+		ctxc.Name = fmt.Sprintf("%s[%d]", name, i)
+		if err := d.decode(ctxc, currentData, currentField); err != nil {
 			errors = appendErrors(errors, err)
 		}
 	}
@@ -874,7 +905,7 @@ func (d *Decoder) decodeArray(name string, data interface{}, val reflect.Value) 
 	return nil
 }
 
-func (d *Decoder) decodeStruct(name string, data interface{}, val reflect.Value) error {
+func (d *Decoder) decodeStruct(ctx *DecodeContext, data interface{}, val reflect.Value) error {
 	dataVal := reflect.Indirect(reflect.ValueOf(data))
 
 	// If the type of the value to write to and the data match directly,
@@ -886,14 +917,14 @@ func (d *Decoder) decodeStruct(name string, data interface{}, val reflect.Value)
 
 	dataValKind := dataVal.Kind()
 	if dataValKind != reflect.Map {
-		return fmt.Errorf("'%s' expected a map, got '%s'", name, dataValKind)
+		return fmt.Errorf("'%s' expected a map, got '%s'", ctx.Name, dataValKind)
 	}
 
 	dataValType := dataVal.Type()
 	if kind := dataValType.Key().Kind(); kind != reflect.String && kind != reflect.Interface {
 		return fmt.Errorf(
 			"'%s' needs a map with string keys, has '%s' keys",
-			name, dataValType.Key().Kind())
+			ctx.Name, dataValType.Key().Kind())
 	}
 
 	dataValKeys := make(map[reflect.Value]struct{})
@@ -953,6 +984,9 @@ func (d *Decoder) decodeStruct(name string, data interface{}, val reflect.Value)
 		}
 	}
 
+	ctxc := ctx.copy()
+	name := ctxc.Name
+
 	// for fieldType, field := range fields {
 	for _, f := range fields {
 		field, fieldValue := f.field, f.val
@@ -1006,11 +1040,13 @@ func (d *Decoder) decodeStruct(name string, data interface{}, val reflect.Value)
 
 		// If the name is empty string, then we're at the root, and we
 		// don't dot-join the fields.
-		if name != "" {
-			fieldName = fmt.Sprintf("%s.%s", name, fieldName)
+		if name == "" {
+			ctxc.Name = fieldName
+		} else {
+			ctxc.Name = fmt.Sprintf("%s.%s", name, fieldName)
 		}
 
-		if err := d.decode(fieldName, rawMapVal.Interface(), fieldValue); err != nil {
+		if err := d.decode(ctxc, rawMapVal.Interface(), fieldValue); err != nil {
 			errors = appendErrors(errors, err)
 		}
 	}