When decoding into a generic map (or a struct that contains a generic map) knowing the source's type information isn't enough information to do type conversions in a hook.

Adding the `name` improves this, as it makes it possible to create a "type converter registry".

Adding a new func type for this is unfortunate, but that is to preserver backwards compability.

```bash
benchmark              old ns/op     new ns/op     delta
Benchmark_Decode-4     6991          7183          +2.75%

benchmark              old allocs     new allocs     delta
Benchmark_Decode-4     51             54             +5.88%

benchmark              old bytes     new bytes     delta
Benchmark_Decode-4     1992          2088          +4.82%
```

Fixes #122